% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.markdown.R
\name{as.markdown}
\alias{as.markdown}
\title{Function to transform statquote objects to strings suitable for markdown}
\usage{
as.markdown(quotes, form = "> *\%s* -- \%s\\n\\n")
}
\arguments{
\item{quotes}{an object of class \code{statquote} returned from functions such as
\code{\link{search_quotes}} or \code{\link{statquote}}}

\item{form}{structure of the markdown output for the text (first argument)
and source (second argument) passed to \code{\link{sprintf}}}
}
\value{
character vector of formatted markdown quotes
}
\description{
This function coerces statquote objects to strings suitable for rendering in markdown.
Quotes and sources are placed within output
formatted via the \code{\link{sprintf}} function.
}
\examples{

ll <- search_quotes("Tukey")
as.markdown(ll)

}
