% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.markdown.R
\name{as.tagged}
\alias{as.tagged}
\title{Transform quotes to tagged key:value pairs}
\usage{
as.tagged(quotes, qid = TRUE)
}
\arguments{
\item{quotes}{an object of class \code{statquote} returned from functions such as
\code{\link{search_quotes}} or \code{\link{statquote}}}

\item{qid}{logical. Should the quote id number `qid` be included in the output?}
}
\value{
A character vector of lines
}
\description{
This function formats a statquote object to the tagged \code{key:value} format used for
maintaining the statquotes database.  The key names are:
\preformatted{
  quo:This is a quotation.
  src:Person or persons who said or wrote the quote.
  cit:Citation for the original quote.
  url:URL where the quote can be found (such as journal articles).
  tag:Comma-separated tags to categorize the quote.
  tex:TeX-formatted citation
}
}
\examples{
qitems <- search_quotes("Yates")
cat(as.tagged(qitems[1:5,]))
}
\seealso{
\code{\link{as.data.frame.statquote}}, \code{\link{as.latex}}, \code{\link{as.markdown}}
}
