% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_quotes.R
\name{expose_quotes}
\alias{expose_quotes}
\title{Expose the quotes data frame}
\usage{
expose_quotes()
}
\description{
In normal use, \code{statquote} hides the quotes data in an environment.
This function makes the quotes data frame visible in the global environment.
}

