% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_quotes.R
\name{find_duplicate_quotes}
\alias{find_duplicate_quotes}
\title{Check for duplicate quotes}
\usage{
find_duplicate_quotes()
}
\description{
Returns a list with aggressively fuzzy matched quotations, along with their
relevant citation information.
}
\examples{

# As the number of quotes has grown, this has become very slow.
# dups <- find_duplicate_quotes()


}
\author{
Phil Chalmers
}
