% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statquote.R
\name{quote_topics}
\alias{quote_topics}
\title{List the topics of the quotes data base}
\usage{
quote_topics(subtopics = FALSE)
}
\arguments{
\item{subtopics}{logical; if \code{TRUE} the subtopics are printed as well
with the associated topic}
}
\description{
List the topics of the quotes data base
}
\examples{
quote_topics()
quote_topics(TRUE)
}
