% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\docType{data}
\name{quotes}
\alias{quotes}
\title{Quotes on statistics, data visualization and science}
\format{
A data frame
}
\usage{
data(quotes)
}
\description{
A data frame with quotations.  The variables are:
}
\details{
\itemize{
  \item \code{qid} quote ID, a numeric vector
  \item \code{text} text of the quote
  \item \code{source} person(s) who said the quote.
  \item \code{citation} citation of the quote
  \item \code{url} URL of the quote
  \item \code{tags} tags used for searching
  \item \code{tex} TeX-style citation
}
}
\keyword{datasets}
