% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quotes_raw.R, R/read_write_quotes.R
\name{read_quotes_raw}
\alias{read_quotes_raw}
\title{Parse quotes from the file quotes_raw.txt.}
\usage{
read_quotes_raw(file = file.path(getwd(), "data-raw/quotes_raw.txt"))

read_quotes_raw(file = file.path(getwd(), "data-raw/quotes_raw.txt"))
}
\arguments{
\item{file}{The file of raw quotes.}
}
\value{
Dataframe with quotes

Dataframe with quotes
}
\description{
There should be no reason for a person to call this function.
This function parses `data-raw/quotes_raw.txt`.
The resulting dataframe is then saved to `data/quotes.rda`.
Although it would be possible to use this function to parse
the quotes when loading the package, that would make it much
slower to load the package.

There should be no reason for a person to call this function.
This function parses `data-raw/quotes_raw.txt`.
The resulting dataframe is then saved to `data/quotes.rda`.
Although it would be possible to use this function to parse
the quotes when loading the package, that would make it much
slower to load the package.
}
