% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_quotes.R
\name{find_duplicate_quotes}
\alias{find_duplicate_quotes}
\title{Check for duplicate quotes}
\usage{
find_duplicate_quotes()
}
\description{
Returns a list with qid, source, and the text where strings are
aggressively fuzzy matched.
}
\examples{

# This can be slow...
# find_duplicate_quotes()


}
\author{
Phil Chalmers
}
