% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_quotes.R
\name{find_duplicate_quotes}
\alias{find_duplicate_quotes}
\title{Check for duplicates in master .csv file}
\usage{
find_duplicate_quotes()
}
\description{
Returns a list with qid, source, and the text where strings are
aggressively fuzzy matched.
}
