% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\docType{data}
\name{quotes}
\alias{quotes}
\title{Quotes on statistics, data visualization and science}
\format{A data frame with 231 rows and 5 variables}
\usage{
data(quotes)
}
\description{
A data frame with over 200 quotations.  The variables are:
}
\details{
\itemize{
  \item \code{qid} quote ID, a numeric vector
  \item \code{topic} main topic, a factor with levels \code{Computing} \code{Data} \code{Data visualization} \code{History} \code{Reviews} \code{Science} \code{Statistics} \code{Unclassified}
  \item \code{subtopic} sub topic, a factor with levels \code{} \code{Averages} \code{Box quotes} \code{Counts} \code{Design} \code{Ellipses} \code{Generalizations} \code{Milestones} \code{Pictures} \code{Tables} \code{Tidy data} \code{Time} \code{Tukey quotes}
  \item \code{text} text of the quote, a character vector
  \item \code{source} source of the quote, a character vector
}
}
\keyword{datasets}

