% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_quotes.R
\name{search_quotes}
\alias{search_quotes}
\title{Function to search quote database}
\usage{
search_quotes(search, fuzzy = FALSE)
}
\arguments{
\item{search}{A character string, used to search the database. Regular
expression characters are allowed.}
}
\value{
A data frame object containing all quotes that match the search
parameters.
}
\description{
This function takes a search pattern (can use regular expressions) and returns all quotes
that match the pattern. If fuzzy is FALSE, then only exact matches are returned (case sensitive).
}
\examples{
search_quotes("^D") # regex to find all quotes that start with "D"
dat <- search_quotes("Tukey") #all quotes with "Tukey"
dat
search_quotes("bad answer", fuzzy = TRUE) # fuzzy match

}
\seealso{
\code{\link{statquote}}, \code{\link{quote_topics}}, \code{\link{quotes}}
}

