% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statquote.R
\name{statquote}
\alias{statquote}
\alias{print.statquote}
\alias{as.data.frame.statquote}
\title{Display a randomly chosen statistical quote}
\usage{
statquote(ind, tag = NULL, source = NULL)

\method{print}{statquote}(x, width = NULL, ...)

\method{as.data.frame}{statquote}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{ind}{Integer vector of quote ID numbers.
If missing, a random value is sampled from all quotations.}

\item{tag}{A character string, used to select a subset of the quotes based
on the tags.}

\item{source}{A character string, used to select a subset of the quotes based
on the source for the quote.}

\item{x}{object of class \code{'statquote'}}

\item{width}{Optional column width parameter}

\item{...}{Other optional arguments}

\item{row.names}{see \code{\link{as.data.frame}}}

\item{optional}{see \code{\link{as.data.frame}}}
}
\value{
A character vector containing one randomly selected quote
from the included data set. It is of class \code{statquote} for
which an S3 print method will be invoked.
}
\description{
This function displays a random statistical quote.
}
\examples{
statquote(10:11)
set.seed(1234)
statquote(123)
statquote(source="Tukey")
statquote(tag="science")
print.data.frame(statquote(301)) # All information, including URL

}
\seealso{
\code{\link{quote_tags}}, \code{\link{search_quotes}}, \code{\link{quotes}},
Inspired by: \code{gaussfact} (\code{https://github.com/eddelbuettel/gaussfacts}),
 \code{\link[fortunes:fortunes]{fortune}}
}
