% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statquote.R
\name{statquote}
\alias{print.statquote}
\alias{statquote}
\title{Function to display a randomly chosen statistical quote}
\usage{
statquote(ind, topic = NULL)

\method{print}{statquote}(x, width = NULL, ...)
}
\arguments{
\item{ind}{Optional index of a quote; if missing a random value is sampled from
the available quotations.}

\item{topic}{A character string, used to select a subset of the quotes based
on the assigned topics.}

\item{x}{Default object for \code{print} method}

\item{width}{Optional column width parameter}

\item{...}{Other optional arguments}
}
\value{
A character vector containing one randomly selected quote
   from the included data set. It is of class \code{statquote} for
   which an S3 print method will be invoked.
}
\description{
This function displays a randomly statistical quote from
a collection. The quotations are classified by topics
}
\examples{
 set.seed(1234)
 statquote()

}
\seealso{
Inspired by: \code{gaussfact} (\code{https://github.com/eddelbuettel/gaussfacts}),
 \code{\link[fortunes:fortunes]{fortune}}
}

